+++ dof_discord_bot/res/restricted_file_handler.py
@@ -1,8 +1,5 @@
 """
-Restricted File Handler
-=======================
-
-Helper module storing the class to load in the logging config files.
+Helper module storing the class to be used within the logging config files.
 """
 import logging as _logging